requirements.txt
streamlit>=1.36
openai>=1.40
numpy==1.26.4
chromadb==0.4.24
pysqlite3-binary>=0.5.2
langchain-text-splitters>=0.0.1
pypdf>=4.2
python-dotenv>=1.0
